Move existing archived items to the trash
Something I hadn't thought about until implementing this was that in
order to migrate things TO the trash, we need the trash to exist -
because foreign keys require that the `collection_id` be an existing
collection.

So we need to create the trash collection in its own migration. We can
delete the code that ensures the trash collection exists, though it's
not strictly necessary. I'll do that in a followup PR.